+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [4.2.6] - 2022-03-16
+### Changed
+- Hide resource table columns without data after loading data.
+### Fixed
+- The use of a number range filter for the Count column type of
+  the ResearchStudy table has been restored.
+
 ## [4.2.5] - 2022-03-08
 ### Fixed
 - Partial hiding of expanded filter lists for research study table.